refactor by creating useAddCategories hook
